requirements.txt
passlib~=1.7.4
pymongo~=4.5.0
flask~=2.3.3
jsonschema~=4.19.0
torchvision~=0.15.2
torch~=2.0.1
pandas~=2.1.4
pyproj~=3.6.1
fiona~=1.9.5
geopandas~=0.14.2
Pillow~=10.3.0
fonttools~=4.51.0
pytest~=7.4.0
google~=3.0.0
flask-cors~=4.0.0
flasgger=0.9.5